package/test_requirements.txt
mock==4.0.3
python-dotenv==0.18.0
pathlib==1.0.1

azure-common==1.1.27
azure-mgmt==5.0.0